+++ RELEASENOTES.md
@@ -6,4 +6,5 @@
 - YamlDotNet 3.0.1 -> 3.1.1
 
 ## Version 1.0.1.27
-- ...
+- fix session flags type casting
+- carIdxSessionFlags returned to